requirements.txt
faiss-cpu
streamlit
torch
clip
Pillow
numpy
